+++ src/main/java/org/fit/ssapp/service/StableMatchingService.java
@@ -20,7 +20,13 @@ import org.fit.ssapp.ss.smt.result.MatchingSolution;
 import org.fit.ssapp.ss.smt.result.MatchingSolutionInsights;
 import org.fit.ssapp.util.ComputerSpecsUtil;
 import org.moeaframework.Executor;
-import org.moeaframework.core.*;
+import org.moeaframework.core.NondominatedPopulation;
+import org.moeaframework.core.Problem;
+import org.moeaframework.core.Problem;
+import org.moeaframework.core.Solution;
+import org.moeaframework.core.TerminationCondition;
+import org.moeaframework.core.Variation;
+import org.moeaframework.core.Variation;
 import org.moeaframework.core.spi.OperatorFactory;
 import org.moeaframework.core.spi.OperatorProvider;
 import org.moeaframework.core.termination.MaxFunctionEvaluations;
@@ -47,6 +53,27 @@ import org.springframework.stereotype.Service;
 @Slf4j
 @RequiredArgsConstructor
 public class StableMatchingService implements ProblemService {
+
+  static {
+    OperatorFactory.getInstance().addProvider(new OperatorProvider() {
+      public String getMutationHint(Problem problem) {
+        return "CustomVariation";
+      }
+
+      public String getVariationHint(Problem problem) {
+        return "CustomVariation";
+      }
+
+      public Variation getVariation(String name, TypedProperties properties, Problem problem) {
+        if (name.equalsIgnoreCase("CustomVariation")) {
+          double crossoverRate = properties.getDouble("CustomVariation.crossoverRate", 0.9);
+          double mutationRate = properties.getDouble("CustomVariation.mutationRate", 0.1);
+          return new CustomVariation(crossoverRate, mutationRate, problem.getNumberOfVariables());
+        }
+        return null;
+      }
+    });
+  }
 
   private static final int RUN_COUNT_PER_ALGORITHM = 10;
   private final SimpMessagingTemplate simpMessagingTemplate;
@@ -75,27 +102,27 @@ public class StableMatchingService implements ProblemService {
       //            log.info("req cap {}, pro cap {}, equality {}", requestAddress, problemAddress,
       //            Objects.equals(requestAddress, problemAddress));
       log.info("Start solving: {}, problem name: {}, problem size: {}",
-              problem.getMatchingTypeName(),
-              problem.getName(),
-              problem.getMatchingData().getSize());
+          problem.getMatchingTypeName(),
+          problem.getName(),
+          problem.getMatchingData().getSize());
       long startTime = System.currentTimeMillis();
 
       NondominatedPopulation results = solveProblem(problem,
-              request.getAlgorithm(),
-              request.getPopulationSize(),
-              request.getGeneration(),
-              request.getMaxTime(),
-              request.getDistributedCores());
+          request.getAlgorithm(),
+          request.getPopulationSize(),
+          request.getGeneration(),
+          request.getMaxTime(),
+          request.getDistributedCores());
 
       if (Objects.isNull(results)) {
         return ResponseEntity
-                .status(HttpStatus.INTERNAL_SERVER_ERROR)
-                .body(Response
-                        .builder()
-                        .status(HttpStatus.INTERNAL_SERVER_ERROR.value())
-                        .message("Error solving stable matching problem.")
-                        .data(null)
-                        .build());
+            .status(HttpStatus.INTERNAL_SERVER_ERROR)
+            .body(Response
+                .builder()
+                .status(HttpStatus.INTERNAL_SERVER_ERROR.value())
+                .message("Error solving stable matching problem.")
+                .data(null)
+                .build());
       }
       //            Testing tester = new Testing((Matches) results.get(0).getAttribute("matches"),
       //            problem.getMatchingData().getSize(), problem.getMatchingData().getCapacities());
@@ -111,30 +138,30 @@ public class StableMatchingService implements ProblemService {
 
       MatchingSolution matchingSolution = formatSolution(algorithm, results, runtime);
       matchingSolution.setSetSatisfactions(problem.getMatchesSatisfactions((Matches) results
-              .get(0)
-              .getAttribute(StableMatchingConst.MATCHES_KEY)));
+          .get(0)
+          .getAttribute(StableMatchingConst.MATCHES_KEY)));
 
       return ResponseEntity.ok(Response
-              .builder()
-              .status(200)
-              .message(
-                      "[Service] Stable Matching: Solve stable matching problem successfully!")
-              .data(matchingSolution)
-              .build());
+          .builder()
+          .status(200)
+          .message(
+              "[Service] Stable Matching: Solve stable matching problem successfully!")
+          .data(matchingSolution)
+          .build());
     } catch (Exception e) {
       log.error("[Service] Stable Matching: Error solving stable matching problem: {}",
-              e.getMessage(),
-              e);
+          e.getMessage(),
+          e);
       // Handle exceptions and return an error response
       return ResponseEntity
-              .status(HttpStatus.INTERNAL_SERVER_ERROR)
-              .body(Response
-                      .builder()
-                      .status(HttpStatus.INTERNAL_SERVER_ERROR.value())
-                      .message(
-                              "[Service] Stable Matching: Error solving stable matching problem.")
-                      .data(null)
-                      .build());
+          .status(HttpStatus.INTERNAL_SERVER_ERROR)
+          .body(Response
+              .builder()
+              .status(HttpStatus.INTERNAL_SERVER_ERROR.value())
+              .message(
+                  "[Service] Stable Matching: Error solving stable matching problem.")
+              .data(null)
+              .build());
     }
   }
 
@@ -199,9 +226,9 @@ public class StableMatchingService implements ProblemService {
             .withMaxEvaluations(generation * populationSize)
             .withTerminationCondition(maxEval)
             .withProperties(properties)
-                .withProperty("operator", "CustomVariation")
-                .withProperty("CustomVariation.crossoverRate", 0.9)
-                .withProperty("CustomVariation.mutationRate", 0.1)
+            .withProperty("operator", "CustomVariation")
+            .withProperty("CustomVariation.crossoverRate", 0.9)
+            .withProperty("CustomVariation.mutationRate", 0.1)
             .distributeOnAllCores()
             .run();
       } else {
@@ -212,9 +239,9 @@ public class StableMatchingService implements ProblemService {
             .withMaxEvaluations(generation * populationSize)
             .withTerminationCondition(maxEval)
             .withProperties(properties)
-                .withProperty("operator", "CustomVariation")
-                .withProperty("CustomVariation.crossoverRate", 0.9)
-                .withProperty("CustomVariation.mutationRate", 0.1)
+            .withProperty("operator", "CustomVariation")
+            .withProperty("CustomVariation.crossoverRate", 0.9)
+            .withProperty("CustomVariation.mutationRate", 0.1)
             .distributeOn(numberOfCores)
             .run();
       }
@@ -236,8 +263,8 @@ public class StableMatchingService implements ProblemService {
                                               String sessionCode) {
     String[] algorithms = StableMatchingConst.ALLOWED_INSIGHT_ALGORITHMS;
     simpMessagingTemplate.convertAndSendToUser(sessionCode,
-            "/progress",
-            createProgressMessage("Initializing the problem..."));
+        "/progress",
+        createProgressMessage("Initializing the problem..."));
     MTMProblem problem = StableMatchingProblemMapper.toMTM(request);
 
     log.info("Start benchmarking {} session code {}", problem.getName(), sessionCode);
@@ -250,8 +277,8 @@ public class StableMatchingService implements ProblemService {
     // of the algorithms
     //        log.info("Start benchmarking the algorithms...");
     simpMessagingTemplate.convertAndSendToUser(sessionCode,
-            "/progress",
-            createProgressMessage("Start benchmarking the algorithms..."));
+        "/progress",
+        createProgressMessage("Start benchmarking the algorithms..."));
 
     for (String algorithm : algorithms) {
       for (int i = 0; i < RUN_COUNT_PER_ALGORITHM; i++) {
@@ -259,11 +286,11 @@ public class StableMatchingService implements ProblemService {
         long start = System.currentTimeMillis();
 
         NondominatedPopulation results = solveProblem(problem,
-                algorithm,
-                request.getGeneration(),
-                request.getPopulationSize(),
-                request.getMaxTime(),
-                request.getDistributedCores());
+            algorithm,
+            request.getGeneration(),
+            request.getPopulationSize(),
+            request.getMaxTime(),
+            request.getDistributedCores());
 
         long end = System.currentTimeMillis();
         assert results != null;
@@ -272,8 +299,8 @@ public class StableMatchingService implements ProblemService {
 
         // send the progress to the client
         String message =
-                "Algorithm " + algorithm + " finished iteration: #" + (i + 1) + "/"
-                        + RUN_COUNT_PER_ALGORITHM;
+            "Algorithm " + algorithm + " finished iteration: #" + (i + 1) + "/"
+                + RUN_COUNT_PER_ALGORITHM;
         Progress progress = createProgress(message, runtime, runCount, maxRunCount);
         System.out.println(progress);
         simpMessagingTemplate.convertAndSendToUser(sessionCode, "/progress", progress);
@@ -287,15 +314,15 @@ public class StableMatchingService implements ProblemService {
     }
     log.info("Benchmark finished! {} session code {}", problem.getName(), sessionCode);
     simpMessagingTemplate.convertAndSendToUser(sessionCode,
-            "/progress",
-            createProgressMessage("Benchmarking finished!"));
+        "/progress",
+        createProgressMessage("Benchmarking finished!"));
 
     return ResponseEntity.ok(Response
-            .builder()
-            .status(200)
-            .message("Get problem result insights successfully!")
-            .data(matchingSolutionInsights)
-            .build());
+        .builder()
+        .status(200)
+        .message("Get problem result insights successfully!")
+        .data(matchingSolutionInsights)
+        .build());
   }
 
   private MatchingSolutionInsights initMatchingSolutionInsights(String[] algorithms) {
@@ -317,12 +344,12 @@ public class StableMatchingService implements ProblemService {
 
   private Progress createProgressMessage(String message) {
     return Progress
-            .builder()
-            .inProgress(
-                    false)
-            // this object is just to send a message to the client, not to show the progress
-            .message(message)
-            .build();
+        .builder()
+        .inProgress(
+            false)
+        // this object is just to send a message to the client, not to show the progress
+        .message(message)
+        .build();
   }
 
   private Progress createProgress(String message,
@@ -331,15 +358,15 @@ public class StableMatchingService implements ProblemService {
                                   int maxRunCount) {
     int percent = runCount * 100 / maxRunCount;
     int minuteLeft = (int) Math.ceil(
-            ((maxRunCount - runCount) * runtime) / 60); // runtime is in seconds
+        ((maxRunCount - runCount) * runtime) / 60); // runtime is in seconds
     return Progress
-            .builder()
-            .inProgress(true) // this object is just to send to the client to show the progress
-            .message(message)
-            .runtime(runtime)
-            .minuteLeft(minuteLeft)
-            .percentage(percent)
-            .build();
+        .builder()
+        .inProgress(true) // this object is just to send to the client to show the progress
+        .message(message)
+        .runtime(runtime)
+        .minuteLeft(minuteLeft)
+        .percentage(percent)
+        .build();
   }
 
   private double getFitnessValue(NondominatedPopulation result) {
